requirements.txt
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.4.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmake==3.27.5
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.14.5
dill==0.3.7
et-xmlfile==1.1.0
filelock==3.13.1
fiona==1.9.5
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.6.0
gekko==1.0.6
geopandas==0.14.0
greenlet==3.0.1
huggingface-hub==0.17.2
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
kor==1.0.0
langchain==0.0.301
langsmith==0.0.40
lit==17.0.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.28.1
openpyxl==3.1.2
optimum==1.13.2
packaging==23.2
pandas==1.5.3
peft==0.5.0
Pillow==10.0.1
psutil==5.9.6
pyarrow==13.0.0
pydantic==2.4.0
pydantic_core==2.10.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
traitlets==5.13.0
transformers==4.33.2
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.7
xxhash==3.3.0
yarl==1.9.2
